Updating the include of some headers that was moved
