+++ src/utils/log/log.c
@@ -1,4 +1,4 @@
-#include "utils/log.h"
+#include "utils/log/log.h"
 
 #ifdef _ENABLE_LOGS_
